Go-based todo cli
This is a GO based CLI todo app to practice basic GO features and play
around with the GO language.
It implements `add`, `list`, `update` features. I plan to finish up on
it. Remaining Features are:
	- mark task as Done(partially implemented)
	- list unfishined tasks
	- add timeline on tasks
	- delete tasks

To make it alittle complex, I plan to:
	- Add database for long term storage
	- Add Authentication to support more users
	- Add task comments for collaboration

Long term Goal for this simple exercise:
	- make it cuter (`UI`)
	- add Graphql APIs
	- hook up a simple web frontend