Simple pub-sub with rabbitMQ
This is a simple Pub-sub module using rabbitMQ. Its running both
publisher and subscriber as child thread and There is a Main thread
acting as the parent to manage these child threads. This makes it easy
to share data in memory between the Publisher and Subscriber.

Its a simple Pub-sub to pass time as I hallucinate on simple complex
systems :)