feat(c-python): This is an example showing how python interfaces with C
There are 2 programs,
- `main.c` which is compiled to the shared `libsum.so`
- `main.py` which is the "exposed" entrypoint to all this magic.

The python code invokes the Cfunction, turning off python GIL and
letting C do it's magic in it's own way and waits for a response.

To make it alittle crazier, I explored the concept of callback functions
to have the wild C logic callback a python function.

Humanity has really nailed this tech thingy.